Split fan_miot.py to vendor-specific integrations
* The main implementation is now under miio/integrations/fan/dmaker/
* The FanZA5 (zhimi) is now under miio/integrations/fan/zhimi/
