+++ miio/__init__.py
@@ -35,11 +35,12 @@ from miio.device import Device, DeviceStatus
 from miio.exceptions import DeviceError, DeviceException
 from miio.fan import Fan, FanP5, FanSA1, FanV2, FanZA1, FanZA4
 from miio.fan_leshow import FanLeshow
-from miio.fan_miot import Fan1C, FanMiot, FanP9, FanP10, FanP11, FanZA5
 from miio.gateway import Gateway
 from miio.heater import Heater
 from miio.heater_miot import HeaterMiot
 from miio.huizuo import Huizuo, HuizuoLampFan, HuizuoLampHeater, HuizuoLampScene
+from miio.integrations.fan.dmaker import Fan1C, FanMiot, FanP9, FanP10, FanP11
+from miio.integrations.fan.zhimi import FanZA5
 from miio.integrations.petwaterdispenser import PetWaterDispenser
 from miio.integrations.vacuum.dreame.dreamevacuum_miot import DreameVacuumMiot
 from miio.integrations.vacuum.mijia import G1Vacuum
