+++ miio/integrations/fan/dmaker/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .fan_miot import Fan1C, FanMiot, FanP9, FanP10, FanP11
