+++ miio/integrations/fan/zhimi/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .zhimi_miot import FanStatusZA5, FanZA5, OperationModeFanZA5
